chat_py_backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.27.1
motor==3.7.1
pymongo==4.13.2
python-jose[cryptography]==3.5.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
python-dotenv==1.1.1
pydantic[email]==2.11.7
pydantic-settings==2.1.0
email-validator==2.2.0
